[refactor] Updated database utility imports for gallery and splat operations; introduced new gallery_tb_utils for improved data fetching
