+++ src/app/lib/db/splat_tb_utils.ts
@@ -2,6 +2,8 @@ import { eq } from "drizzle-orm";
 import { SplatEditMetaData, SplatUploadMetaData } from "../definitions/SplatPayload";
 import { db } from "./db";
 import { splats } from "./schema";
+import SceneItem from "../definitions/SceneItem";
+import VideoItem from "../definitions/VideoItem";
 
 export async function insertNewRowInDB(splatUploadMetaData: SplatUploadMetaData): Promise<number | null> {
   const ids = await db
@@ -18,6 +20,54 @@ export async function insertNewRowInDB(splatUploadMetaData: SplatUploadMetaData)
     return ids[0].insertedId;
   }
   return null;
+}
+
+export async function fetchVideoItems(): Promise<VideoItem[]> {
+  const combinedData = await db
+    .select({
+      id: splats.id,
+      name: splats.name,
+      video: splats.video,
+      splat: splats.splat,
+    })
+    .from(splats)
+    .orderBy(splats.id)
+    .then((data) => {
+      return data
+        .map((item) => ({
+          id: item.id,
+          name: item.name || "",
+          src: item.video || "",
+          splatUrl: item.splat || "",
+        }))
+        .sort((a, b) => a.id - b.id);
+    });
+
+  return combinedData;
+}
+
+export async function getSceneItemById(id: number): Promise<SceneItem | null> {
+  const sceneItems: SceneItem[] = await db
+    .select({
+      id: splats.id,
+      name: splats.name,
+      description: splats.description,
+      splatUrl: splats.splat,
+      videoUrl: splats.video
+    })
+    .from(splats)
+    .where(eq(splats.id, id))
+    .then((data: SceneItem[]) => {
+      return data.map((item: SceneItem) => ({
+        id: item.id,
+        name: item.name || "",
+        description: item.description || "",
+        splatUrl: item.splatUrl || "",
+        videoUrl: item.videoUrl || ""
+      }));
+    });
+
+  return sceneItems.length ? sceneItems[0] : null;
 }
 
 export async function updateRowWithID(splatEditMetaData: SplatEditMetaData): Promise<number | null> {
@@ -62,3 +112,7 @@ export async function deleteRowWithID(id: number): Promise<number | null> {
       return null;
     });
 }
+
+// export async function getSplatUrlWithId(id: number): Promise <string | null> {
+//   return db.select(splats).where(eq(splats.id, id)).returning({splatUrl: splats.splat}).then(())
+// }
