+++ src/app/lib/definitions/GalleryItem.ts
@@ -1,6 +1,17 @@
-export default interface GalleryItem {
+export interface GalleryItem {
     id: number;
     name: string;
     description: string | null,
     thumbnailUrl: string | null;
 }
+export interface GalleryDetails {
+    name: string;
+    description: string | null;
+  }
+
+  export type GallerySplat = {
+    id: number;
+    video: string | null;
+    splat: string | null;
+  }
+
